requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.3
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
isort==5.13.2
jiter==0.6.1
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.2
mypy-extensions==1.0.0
numpy==2.0.2
openai==1.52.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pygltflib==1.16.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tomli==2.2.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.17.0
